requirements.txt
requests
transformers
python-dotenv
torch
